ml-service/api-azure/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2